+++ server-worker/importers/regions/nz-chc.js
@@ -1,4 +1,4 @@
-const defaultConfig = require('./defaultConfig.js')
+const defaultConfig = require('../defaultConfig.js')
 
 const christchurch = defaultConfig.get('metro-christchurch', {
   url: 'http://metroinfo.co.nz/Documents/gtfs.zip',
