+++ server-worker/importers/regions/nz-otg.js
@@ -1,4 +1,4 @@
-const defaultConfig = require('./defaultConfig.js')
+const defaultConfig = require('../defaultConfig.js')
 
 const otago = defaultConfig.get('otago', {
   url: 'https://www.orc.govt.nz/transit/google_transit.zip',
